requirements.txt
Flask_JWT_Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-Marshmallow
SQLAlchemy
Werkzeug
uwsgi
psycopg2
redis